Compiling infrastructure - Moving some functionality #8
