+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jsoncv/core",
-  "version": "0.2.1",
+  "version": "0.2.2",
   "description": "JSON CV's Core Functionalities",
   "license": "MIT",
   "main": "lib/index.js",
