+++ src/index.ts
@@ -1,7 +1,9 @@
 import validator from './validator'
 import server from './server'
+import { exportToHtml } from './export'
 
 export {
     validator,
-    server
+    server,
+    exportToHtml
 }
