+++ src/utils.ts
@@ -1,6 +1,7 @@
 import { cwd } from 'process'
 import axios from 'axios'
 import fs from 'fs'
+const requireg = require('requireg')
 
 // Checking to see if a string is a valid URL
 const isUrlRegex = new RegExp(/https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)/gi);
@@ -28,8 +29,28 @@ const loadFile = (location:string):any => {
     })
 }
 
+// Checking availability of the globally installed theme by name
+const checkIfThemeIsGloballyAvailable = (name:string):boolean => {
+    return requireg.resolve(name) !== undefined
+}
+
+// Getting the location of the globally installed theme by name
+const getGlobalThemeLocation = (name:string):string => {
+    const location = requireg.resolve(name).toString().trim()
+    return location.substring(0, location.indexOf("/index.js"))
+}
+
+// Dynamically loading the theme as a module
+const loadTemplateModule = async (templateLocation:string) => {
+    const absoluteLocation = isLocationAbsolute(templateLocation) ? `${templateLocation}/index.js` : `${cwd()}/${templateLocation}/index.js`
+    return import(absoluteLocation)
+}
+
 export {
     isLocationAbsolute,
     isUrl,
-    loadFile
+    loadFile,
+    checkIfThemeIsGloballyAvailable,
+    getGlobalThemeLocation,
+    loadTemplateModule
 }
